// pop-client/src/app/shared/openapi/model/project-status-choices.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ProjectStatusChoices = 'planned' | 'ongoing' | 'completed' | 'aborted';

export const ProjectStatusChoices = {

    Planned: 'planned' as ProjectStatusChoices,

    Ongoing: 'ongoing' as ProjectStatusChoices,

    Completed: 'completed' as ProjectStatusChoices,

    Aborted: 'aborted' as ProjectStatusChoices
};